+++ src/constants.js
@@ -157,10 +157,12 @@ export const CONTROLLER_EVENT = {
     toggleMode: 'togglemode',
     enter: 'annotationmodeenter',
     exit: 'annotationmodeexit',
+    createThread: 'createannotationthread',
     register: 'registerthread',
     unregister: 'unregisterthread',
     bindDOMListeners: 'binddomlisteners',
-    unbindDOMListeners: 'unbinddomlisteners'
+    unbindDOMListeners: 'unbinddomlisteners',
+    renderPage: 'annotationsrenderpage'
 };
 
 export const CREATE_EVENT = {
@@ -178,5 +180,5 @@ export const ID_MOBILE_ANNOTATION_DIALOG = 'mobile-annotation-dialog';
 
 export const DRAW_RENDER_THRESHOLD = 16.67; // 60 FPS target using 16.667ms/frame
 export const DRAW_BASE_LINE_WIDTH = 3;
-export const DRAW_BORDER_OFFSET = 5;
+export const BORDER_OFFSET = 5;
 export const DRAW_DASHED_SPACING = 5;