+++ src/controllers/DrawingModeController.js
@@ -1,4 +1,3 @@
-import rbush from 'rbush';
 import AnnotationModeController from './AnnotationModeController';
 import shell from './../shell.html';
 import DocDrawingThread from '../doc/DocDrawingThread';
@@ -14,7 +13,6 @@ import {
     SELECTOR_ANNOTATION_DRAWING_HEADER,
     CLASS_ANNNOTATION_DRAWING_BACKGROUND,
     CLASS_ANNOTATION_LAYER_DRAW,
-    DRAW_BORDER_OFFSET,
     CLASS_ACTIVE,
     CLASS_ANNOTATION_MODE,
     CONTROLLER_EVENT
@@ -98,55 +96,6 @@ class DrawingModeController extends AnnotationModeController {
     }
 
     /**
-     * Register a thread that has been assigned a location with the controller
-     *
-     * @inheritdoc
-     * @public
-     * @param {AnnotationThread} thread - The thread to register with the controller
-     * @return {void}
-     */
-    registerThread(thread) {
-        if (!thread || !thread.location) {
-            return;
-        }
-
-        const page = thread.location.page || 1; // Defaults to page 1 if thread has no page'
-        if (!(page in this.threads)) {
-            /* eslint-disable new-cap */
-            this.threads[page] = new rbush();
-            /* eslint-enable new-cap */
-        }
-        this.threads[page].insert(thread);
-        this.emit(CONTROLLER_EVENT.register, thread);
-        thread.addListener('threadevent', (data) => {
-            this.handleThreadEvents(thread, data);
-        });
-    }
-
-    /**
-     * Unregister a previously registered thread that has been assigned a location
-     *
-     * @inheritdoc
-     * @public
-     * @param {AnnotationThread} thread - The thread to unregister with the controller
-     * @return {void}
-     */
-    unregisterThread(thread) {
-        if (!thread || !thread.location) {
-            return;
-        }
-
-        const page = thread.location.page || 1; // Defaults to page 1 if thread has no page'
-        if (!this.threads[page]) {
-            return;
-        }
-
-        this.threads[page].remove(thread);
-        this.emit(CONTROLLER_EVENT.unregister, thread);
-        thread.removeListener('threadevent', this.handleThreadEvents);
-    }
-
-    /**
      * Bind the DOM listeners for this mode
      *
      * @inheritdoc
@@ -221,7 +170,8 @@ class DrawingModeController extends AnnotationModeController {
         const threadParams = this.annotator.getThreadParams([], {}, TYPES.draw);
         this.currentThread = new DocDrawingThread(threadParams);
         this.currentThread.addListener('threadevent', (data) => {
-            this.handleThreadEvents(this.currentThread, data);
+            const thread = this.currentThread || this.selectedThread;
+            this.handleThreadEvents(thread, data);
         });
 
         // Get handlers
@@ -249,7 +199,7 @@ class DrawingModeController extends AnnotationModeController {
         });
 
         this.pushElementHandler(this.postButtonEl, 'click', () => {
-            this.currentThread.saveAnnotation(TYPES.draw);
+            this.saveThread(this.currentThread);
             this.toggleMode();
         });
 
@@ -262,31 +212,28 @@ class DrawingModeController extends AnnotationModeController {
      *
      * @inheritdoc
      * @protected
-     * @param {AnnotationThread} thread - The thread that emitted the event
-     * @param {Object} data - Extra data related to the annotation event
+     * @param {AnnotationThread} thread The thread that emitted the event
+     * @param {Object} data Extra data related to the annotation event
      * @return {void}
      */
     handleThreadEvents(thread, data = {}) {
         const { eventData } = data;
         switch (data.event) {
             case 'softcommit':
-                // Save the original thread, create a new thread and
-                // start drawing at the location indicating the page change
                 this.currentThread = undefined;
-                thread.saveAnnotation(TYPES.draw);
-                this.registerThread(thread);
-
-                this.unbindListeners();
-                this.bindListeners();
+                this.saveThread(thread);
 
                 // Given a location (page change) start drawing at the provided location
                 if (eventData && eventData.location) {
+                    this.unbindListeners();
+                    this.bindListeners();
+
                     this.currentThread.handleStart(eventData.location);
                 }
 
                 break;
             case 'dialogdelete':
-                if (!thread.dialog) {
+                if (!thread || !thread.dialog) {
                     return;
                 }
 
@@ -316,7 +263,7 @@ class DrawingModeController extends AnnotationModeController {
     }
 
     /**
-     * Find the selected drawing threads given a pointer event. Randomly picks one if multiple drawings overlap
+     * Selects a drawing thread given a pointer event. Randomly picks one if multiple drawings overlap
      *
      * @protected
      * @param {Event} event - The event object containing the pointer information
@@ -328,20 +275,10 @@ class DrawingModeController extends AnnotationModeController {
             return;
         }
 
-        const location = this.annotator.getLocationFromEvent(event, TYPES.point);
-        if (!location || Object.keys(this.threads).length === 0 || !this.threads[location.page]) {
-            return;
-        }
-
-        const eventBoundary = {
-            minX: +location.x - DRAW_BORDER_OFFSET,
-            minY: +location.y - DRAW_BORDER_OFFSET,
-            maxX: +location.x + DRAW_BORDER_OFFSET,
-            maxY: +location.y + DRAW_BORDER_OFFSET
-        };
+        event.stopPropagation();
 
         // Get the threads that correspond to the point that was clicked on
-        const intersectingThreads = this.threads[location.page].search(eventBoundary);
+        const intersectingThreads = this.getIntersectingThreads(event);
 
         // Clear boundary on previously selected thread
         this.removeSelection();
@@ -430,6 +367,22 @@ class DrawingModeController extends AnnotationModeController {
                 util.disableElement(this.redoButtonEl);
             }
         }
+    }
+
+    /**
+     * Save the original thread and create a new thread
+     *
+     * @private
+     * @param {AnnotationThread} thread The thread that emitted the event
+     * @return {void}
+     */
+    saveThread(thread) {
+        if (!thread) {
+            return;
+        }
+
+        thread.saveAnnotation(TYPES.draw);
+        this.registerThread(thread);
     }
 }